refactor: Reorganize icon components and improve import structure
- Move AdaptiveIcon to icons directory for better organization
- Update AdaptiveIcon to use default export for consistency
- Consolidate icon exports through icons/index.ts barrel file
- Update imports across navbar, footer, and social-links components
- Remove individual icon exports from main components index
