+++ src/components/index.ts
@@ -6,13 +6,9 @@ export { ThemeToggle } from './theme/theme-toggle';
 
 // UI Components
 export { Button } from './ui/button';
-export { GitHubIcon } from './icons/socials/GitHubIcon';
-export { LinkedInIcon } from './icons/socials/LinkedInIcon';
-export { TwitterIcon } from './icons/socials/TwitterIcon';
 export { SocialLinks } from './social-links';
 export { WordRotate } from './magicui/word-rotate';
 
 // Utility Components
-export { AdaptiveIcon } from './adaptive-icon';
 export { DynamicThemeColor } from './theme/dynamic-theme-color';
 export { DynamicViewport } from './dynamic-viewport';
